Remove min/max bounds on bazel version in `configure.py`.
These checks were useful when we used to set the bazel version in the CI scripts but now that we have `.bazelversion` synced in the repo any user using Bazelisk or a recent version of Bazel will automatically get the exact Bazel version that we are using. Hence, this step in `configure.py` is no longer needed.

Furthermore, we are working on completely removing the need to run `configure.py`, at least for the vast majority of users.

PiperOrigin-RevId: 448033791