Fix profiler segfault when compiling with C++17.
The second argument of `TraceMeEncode` is an initializer list of `TraceMeArg`s. Each `TraceMeArg` is a key-value pair where the value is an `absl::AlphaNum` object (that is, a number of an `absl::string_view`). In C++17, `absl::string_view` aliases to `std::string_view`.

The problem arises when the value is a null pointer. The `absl::string_view` constructor allows building from a `nullptr` value[1], but the C++17 standard does not allow `nullptr` as argument to `std::string_view`[2].

In fact, the issue is slightly worse here, as we are also relying on implicit conversions from `char*` to `string_view`. So we cannot just replace the `string_view` constructor with `absl::NullSafeStringView`

[1]: https://github.com/abseil/abseil-cpp/issues/73
[2]: https://timsong-cpp.github.io/cppwp/string.view.cons

PiperOrigin-RevId: 448997842