Take `_XlaMustCompile` into account in `XlaClusterFormationPass`
Previously, we checked for `_XlaMustCompile` only in `HasQualifiedNonTPUOp` but
not in `XlaClusterFormationPass` which is a bug if there are multiple
partitioned calls with inconsistent `_XlaMustCompile` attributes.
Now we check the attribute at both places which also requires reordering
`XlaClusterFormationPass` and `CanonicalizerPass` because canonicalizer might
lose attributes before we can evaluate them in cluster formation.

PiperOrigin-RevId: 449393040