[grappler][tfg] Make non-import errors fatal
Most users have Grappler configured to soft fail, meaning that if an optimizer fails, it will do so quietly. This is problematic for multiple reasons, but it makes it difficult to catch bugs in TFG passes. Changing the default option to always hard fail is not viable.

Instead, make the Grappler optimizer always hard-fail on the TFG optimizer. However, importer errors will still soft-fail by returning `Aborted`, as is Grappler's convention. This can be due to users supplying invalid graphs or graphs containing unregistered operations in non-generic functions.

That is, any failure after the importer in post-import passes, the pass pipeline, pre-export passes, and exporter, is considered fatal and indicates a bug in TFG code.

PiperOrigin-RevId: 449520161