[mhlo] Add PartitionId Op
Adds `mhlo.partition_id`, corresponding to XLA HLO `kPartitionId` opcode, and
`lmhlo.partition_id`. This is currently only for use in MHLO <-> HLO
roundtrips, so lowering to lmhlo.parition_id is not currently supported. Use of
this op in bufferized MHLO should replace the use of the LMHLO op when we move
from using LMHLO to bufferized MHLO.

PiperOrigin-RevId: 450601384