[mhlo] Make MHLO specific about the bit widths of floating-point types.
Before this change, MHLO uses floating point type (`AnyFloat`) irrespective of the bit-widths which enable MHLO to include bit widths like 80/128 which are not currently exported to HLO (as the later does not support those types). The CL makes the MHLO floating-point types to use one of fp16/bfloat16/fp32/fp64.

PiperOrigin-RevId: 452450147