Add shape inference pass to Remapper pass pipeline
Also add few passes which will turn the IR into canonical form which is better for optimization.
Remove a TODO which was supposed to add several passes by default. Given that it's still in the early stage of moving to TFG, it'll be better to add them to each individual pass pipeline at this moment.

PiperOrigin-RevId: 453560166