[mhlo] Enhance the tests of mhlo.abs.
This op is a simple unary element-wise Op except that it supports complex->real. That why the existing code is using a separated inferReturnTypes() other than `INFER_RETURN_TYPE_COMPONENTS_FROM_OPERANDS`. So add a test to cover the complex case.

Trait `HLO_UnaryElementwiseOp` ensures the return shape (dims) match, so no need of a verbose test about that.

PiperOrigin-RevId: 453778165