Update test_util.py to evaluate CompositeTensors with .numpy method.
This CL prepares for making ResourceVariable CompositeTensor. Once ResourceVariable becomes a CompositeTensor, the tf.map_structure leaf of a ResourceVariable will be the resource handle instead of the ResourceVariable itself. To be able to convert a CompositeTensor with ResourceVariable Components, try the .numpy() method before recursively calling this method on every leaves.

PiperOrigin-RevId: 455647667