Fix ragged reduce_variance for small deviations.
The op uses a less-numerically stable algorithm, which can sometimes
produce negative variances when inputs are all equal (or nearly equal).
Ideally we would subtract the mean from the input, but currently
ragged tensors do not allow broadcasting the ragged dimension if the
size does not match, so we resort to simply ensuring the variance
is non-negative.

Fixes #56222.

PiperOrigin-RevId: 456570443