[mhlo] Implement `mhlo.domain`, exporter, importer
This mirrors the HLO kDomain. Currently, the only type of metadata supported is
ShardingMetadata, because that is the only type of metatdata supported by the
XlaBuilder API we use to export HLO. (By contrast, in XLA, DomainMetadata is
open-world.)

Thankfully, the only other implementation of DomainMetadata is OpNameMetadata,
which is generally used for debugging and never for compiling production
models.

PiperOrigin-RevId: 456869822