Fix interleave_test in CGroup restricted CPU envs
Use `len(os.sched_getaffinity(0))`, which returns the number cores available to the job,
rather than `multiprocessing.cpu_count()`, which returns the total number of cores in the node,
and which causes the interleave_test to fail when e.g. CGroups limit the
number of cores to less than the total core count.