[XLA] Add functionality to split complicated reshardings into smaller pieces that can be more easily handled.
Resharding works by pattern matching known reshard patterns and applying rules to perform the resharding.
Sometimes some reshards can be too complicated to handle, in which case we give up
and use the fallback behavior of rematerializing the entire buffer and then cutting it the way we want.
Sometimes though by doing small steps, like splitting a dimension,
or rematerializing/dematerializing a dimension that was going to/is partial replicated
we can move the situation in a supported case. Adding some functionality match some of these potential cases,
perform a simplifying action and retry calling the Reshard infrastructure again. If we are able to avoid full
rematerialization keep the result, otherwise fallback to the default behavior like before.

PiperOrigin-RevId: 459855012