Disable Eigen's new AVX512 gemm kernels again.
A recent addition from intel into Eigen (!972)
added new specialized GEMM kernels for AVX512. Unfortunately, these conflict with and are
incompatible with specializations in `eigen_spatial_convolutions.h` in TensorFlow.
Here we put in a hack to disable them.

The last attempt to fix this failed, since other headers are sometimes included
before this the `eigen_spatial_convolutions.h` one, which end up including
the conflicting specialized gemm kernel before we have the chance to disable it.
The only option is to disable via `defines` in the BUILD file so that all dependent
targets inherit this option.

See issue #56731.

PiperOrigin-RevId: 460533762