Add eager/graph mode logging message to test_util.run_in_graph_and_eager_modes.
test_util.run_in_graph_and_eager_modes runs a test method first in graph mode then in eager mode. When a test failed, it is hard to tell if it was running in eager or graph mode. This is even more confusing when a test fails both in eager and graph mode but with different reasons.

Ideally we may want to create two test methods with different suffixes like parameterized. This quick change serves as a temp solution.

PiperOrigin-RevId: 461754264