Consolidate handling of CustomCall with computation into CustomCallInternal.
This CL refactors the existing XlaBuilder::CustomCallInternal method to now contain the XlaComputation* arg to be handled as well. This is just setup to a followup CL that will handle the called computation logic in MlirHloBuilder::CustomCallInternal for legalization to MLIR. Keeping that functionality out of this CL to narrow review scope.

PiperOrigin-RevId: 461981355